// src/app/types/enums/enums.ts
export enum TechnologyStack {
  React = 0,
  Angular = 1,
  Vue = 2,
  DotNet = 3,
  Python = 4,
  Java = 5,
  JavaScript = 6,
  TypeScript = 7,
  PHP = 8,
  Ruby = 9,
  Go = 10,
  Rust = 11,
  Swift = 12,
  Kotlin = 13
}

export enum ExperienceLevel {
  Junior = 0,
  Middle = 1,
  Senior = 2
}

export enum ZodiacSign {
  Aries = 0,
  Taurus = 1,
  Gemini = 2,
  Cancer = 3,
  Leo = 4,
  Virgo = 5,
  Libra = 6,
  Scorpio = 7,
  Sagittarius = 8,
  Capricorn = 9,
  Aquarius = 10,
  Pisces = 11
}